requirements.txt
apache-airflow
kafka-python
requests
pandas
sqlalchemy
